+++ packages/client/package.json
@@ -20,11 +20,6 @@
         "import": "./dist/index.mjs",
         "default": "./dist/index.mjs"
       },
-      "./openapi": {
-        "types": "./dist/openapi/index.d.mts",
-        "import": "./dist/openapi/index.mjs",
-        "default": "./dist/openapi/index.mjs"
-      },
       "./rpc": {
         "types": "./dist/rpc/index.d.mts",
         "import": "./dist/rpc/index.mjs",
@@ -39,7 +34,6 @@
   },
   "exports": {
     ".": "./src/index.ts",
-    "./openapi": "./src/openapi/index.ts",
     "./rpc": "./src/rpc/index.ts",
     "./fetch": "./src/adapters/fetch/index.ts"
   },
