--- packages/client/src/openapi/index.ts
@@ -1,3 +1,0 @@
-export * from './bracket-notation'
-export * from './json-serializer'
-export * from './serializer'
