+++ packages/openapi-client/src/adapters/standard/bracket-notation.test.ts
