+++ packages/openapi-client/src/adapters/standard/openapi-json-serializer.ts
