+++ packages/openapi-client/tsconfig.json
@@ -7,6 +7,7 @@
   "include": ["src"],
   "exclude": [
     "**/*.test.*",
+    "**/*.bench.*",
     "**/*.test-d.ts",
     "**/__tests__/**",
     "**/__mocks__/**",
