+++ packages/openapi/src/openapi-generator.ts
@@ -4,8 +4,8 @@ import type { PublicOpenAPIPathParser } from './openapi-path-parser'
 import type { JSONSchema } from './schema'
 import type { SchemaConverter } from './schema-converter'
 import { fallbackORPCErrorStatus } from '@orpc/client'
-import { OpenAPIJsonSerializer } from '@orpc/client/openapi'
 import { type ContractRouter, fallbackContractConfig, getEventIteratorSchemaDetails } from '@orpc/contract'
+import { OpenAPIJsonSerializer } from '@orpc/openapi-client/standard'
 import { type AnyRouter, eachAllContractProcedure } from '@orpc/server'
 import { group } from '@orpc/shared'
 import { type OpenAPI, OpenApiBuilder } from './openapi'
