+++ packages/openapi/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig.lib.json",
   "references": [
     { "path": "../client" },
+    { "path": "../openapi-client" },
     { "path": "../contract" },
     { "path": "../server" },
     { "path": "../standard-server" },
